requirements.txt
Django==5.1.4
djangorestframework==3.15.2
psycopg2==2.9.10
python-dotenv==1.0.1
black==24.10.0
isort==5.13.2
pillow==11.0.0
djangorestframework_simplejwt==5.4.0
django-filter==24.3
coverage==7.6.10
drf-yasg==1.21.8
stripe==11.4.1
requests==2.32.3
redis==5.2.1
celery==5.4.0
eventlet==0.39.0
django-celery-beat==2.7.0